+++ lib/basic_logger_file.dart
@@ -4,10 +4,9 @@ library basic_logger_file;
 
 import 'dart:async';
 import 'dart:io';
+import 'package:basic_logger/basic_logger.dart';
 import 'package:path/path.dart' as path;
 import 'package:logging/logging.dart';
-
-import 'package:basic_logger/basic_logger.dart';
 
 part 'src/basic_logger_file_base.dart';
 // export 'src/basic_logger_file_base.dart';