+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.2
+
+- The logger name is set to selfname.
+- selfonly, when true, filter by selfname. Otherwise, output all.
+
 ## 0.0.1
 
 - Initial version.